requirements.txt
google-cloud-firestore>=2.3.4
google-cloud-storage>=1.44
librosa>=0.8.1
numba>=0.54
numpy<=1.20
scikit-learn>=1.0.2
